Changes to be committed: 	modified:   README.md 	new file:   lib/images/icon_enter@2x.png 	new file:   lib/images/icon_enter@3x.png 	new file:   lib/images/icon_space@2x.png 	new file:   lib/images/icon_space@3x.png 	modified:   lib/index.js
I created a new props for the package, number-pad-comma.
This layout add one more row on component and change the last line.
With this layout after number we see : dot(.) Zero(0) and comma(,) / backspace, space and return.
I've be these necessity in one project and i thought it would be good put this in the main project.